+++ packages/ui/src/index.ts
@@ -1,1 +1,2 @@
-export {};
+export { Layout } from './layout/Layout'
+export type { LayoutProps } from './layout/Layout'